api/requirements.txt
# ============== API REST CORE ==============
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ============== HTTP & REQUESTS ==============
requests==2.31.0
httpx==0.25.2

# ============== CONFIGURATION ==============
python-dotenv==1.0.0

# ============== VALIDATION & TYPES ==============
typing-extensions==4.8.0

# ============== LOGGING ==============
# (logging est inclus dans Python standard)

# ============== OPTIONNEL: CACHE REDIS ==============
# redis==5.0.1
# Décommentez si vous voulez utiliser Redis pour le cache

# ============== OPTIONNEL: MONITORING ==============
# prometheus-client==0.19.0
# Décommentez pour métriques Prometheus

# ============== SÉCURITÉ ==============
cryptography==41.0.7

# ============== DÉVELOPPEMENT (optionnel) ==============
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0

# ============== PRODUCTION (optionnel) ==============
# gunicorn==21.2.0
# Décommentez pour déploiement production avec Gunicorn